+++ CHANGELOG.md
@@ -8,6 +8,7 @@ We use [semantic versioning](http://semver.org/):
 - [feature] _agent_: Experimental support for Java 25 class files
 - [fix] _agent_: Reported wrong version number in docker container
 - [fix] _agent_: Erroneous log messages were printed about unmatched probes
+- [fix] _agent_: Couldn't receive impacted tests when revision is provided instead of commit
 
 # 35.0.3
 - [fix] _teamscale-gradle-plugin_: `IllegalArgumentException`: Can't execute request without discovering it first